[chore]#104 rename 반영
